refactor: 수익을 계산할 때 다운캐스팅으로 인한 오버플로우 발생, 반환형을 double로 변환

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -11,9 +11,9 @@ public enum LottoPrize {
     ETC(0, 0);
 
     private final int prizeRank;
-    private final int prizeMoney;
+    private final double prizeMoney;
 
-    LottoPrize(int matchCount, int prizeMoney) {
+    LottoPrize(int matchCount, double prizeMoney) {
         this.prizeRank = matchCount;
         this.prizeMoney = prizeMoney;
     }
@@ -22,7 +22,7 @@ public enum LottoPrize {
         return prizeRank;
     }
 
-    public int getPrizeMoney() {
+    public double getPrizeMoney() {
         return prizeMoney;
     }
